Setting schedule type for coalition.

+++ src/main/java/ro/ieat/soso/reasoning/CoalitionReasoner.java
@@ -52,6 +52,7 @@ public class CoalitionReasoner {
     TaskUsageMappingRepository taskUsageMappingRepository;
     @Autowired
     ScheduledRepository scheduledRepository;
+    private Double gamma = 0.8;
 
 
     @RequestMapping(method = RequestMethod.GET, path = "/coalitions/init/{time}")
@@ -142,8 +143,10 @@ public class CoalitionReasoner {
                     }
                 }
             }
-            LOG.info(String.format("id: %d\nsize: %d\neta: %d\nmin_cpu: %.4f\nduplicates: %d\n",
-                    c.getId(), c.getMachines().size(), c.getCurrentETA(), mincpu, duplicates));
+            if(duplicates > 0) {
+                LOG.info(String.format("id: %d\nsize: %d\neta: %d\nmin_cpu: %.4f\nduplicates: %d\n",
+                        c.getId(), c.getMachines().size(), c.getCurrentETA(), mincpu, duplicates));
+            }
         }
     }
 
@@ -268,9 +271,12 @@ public class CoalitionReasoner {
             List<Coalition> coalitionList = coalitionRepository.findByScheduleClass(component);
             if(coalitionList.size() > 0) {
                 coalitionList.stream().filter(c -> !scheduledCoalitions.contains(c.getId())).forEach(c -> {
-                    coalitionClient.deleteCoalitionFromComponent(c, component);
-                    toDelete.add(c);
-                    toReorganize.addAll(c.getMachines());
+                    c.setConfidenceLevel(c.getConfidenceLevel() * gamma);
+                    if(c.getConfidenceLevel() < 0.5) {
+                        coalitionClient.deleteCoalitionFromComponent(c, component);
+                        toDelete.add(c);
+                        toReorganize.addAll(c.getMachines());
+                    }
                 });
             }
